Fix eval() usage:
The eval() would mask the error because it wasn't reported. Instead,
we now check that both the eval succeeded and then that it's the
proper result.

The test for it (in e0b1895ea743783299ad63861812d27f7bd161ad) makes
changes to the code for the sake of testing which I would prefer
not to merge.